backend/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
datasets==3.3.2
dill==0.3.8
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.11
filelock==3.17.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.1
grpcio==1.70.0
grpcio-status==1.70.0
httplib2==0.22.0
huggingface-hub==0.29.2
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
mediapipe==0.10.21
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
oauth2==1.9.0.post1
opencv-contrib-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
pyarrow==19.0.1
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
scikit-learn==1.6.1
scipy==1.13.1
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
SQLAlchemy==2.0.38
starlette==0.46.0
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.6.0
torchaudio==2.6.0
torchsummary==1.5.1
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
